podcast_scraper/requirements.txt
yt-dlp==2024.11.18
feedparser==6.0.11
beautifulsoup4==4.12.3
anthropic==0.39.0
ollama==0.5.1
pandas==2.2.2
python-dotenv==1.0.1
tqdm==4.66.5
tenacity==9.0.0
pyyaml==6.0.2
openpyxl==3.1.5
pydantic==2.11.7
requests==2.32.4
